Added InputValidator

+++ MainScreen.cs
@@ -51,8 +51,14 @@ namespace InventoryMgmt
 
         private void PartDelete_Click(object sender, EventArgs e)
         {
-            Part selectedPart = (Part)PartsGridView.SelectedRows[0].DataBoundItem;
-            Inventory.Instance.DeletePart(selectedPart);
+            var result = MessageBox.Show("Are you sure you want to delete this part?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (result == DialogResult.Yes)
+            {
+                //need to check if part is associated with any product
+                Part selectedPart = (Part)PartsGridView.SelectedRows[0].DataBoundItem;
+                Inventory.Instance.DeletePart(selectedPart);
+
+            }
         }
 
         private void ProductAdd_Click(object sender, EventArgs e)
@@ -60,6 +66,16 @@ namespace InventoryMgmt
             AddProduct addproduct = new AddProduct();
             addproduct.ShowDialog();
             ProductsGridView.Refresh();
+        }
+
+        private void ProductDelete_Click(object sender, EventArgs e)
+        {
+            var result = MessageBox.Show("Are you sure you want to delete this Product?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (result == DialogResult.Yes)
+            {
+                Part selectedPart = (Part)PartsGridView.SelectedRows[0].DataBoundItem;
+                Inventory.Instance.DeletePart(selectedPart);
+            }
         }
     }
 }
